UI Improvements, Message Attachments, and Other Changes

+++ src/main/java/UI/ListEmails.java
@@ -1,7 +1,6 @@
 package UI;
 
 import Filter.Mailbox;
-import Message.MessageParser;
 import JSON.WriteMessageStatus;
 import JSON.ReadMessageStatus;
 
@@ -17,6 +16,8 @@ import java.util.stream.Collectors;
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
 import org.json.simple.parser.JSONParser;
+
+import Email.EmailParser;
 
 public class ListEmails extends MainMenu {
     private final int PAGE_SIZE = 10;
@@ -37,7 +38,6 @@ public class ListEmails extends MainMenu {
     ReadMessageStatus readMessageStatus = new ReadMessageStatus();
     JSONObject messagesID = readMessageStatus.readStatus();
     JSONObject messageList = null;
-
 
     public ListEmails(Mailbox mailbox, InputHandler inputHandler) {
         this.mailbox = mailbox;
@@ -107,13 +107,8 @@ public class ListEmails extends MainMenu {
         int start = currentPage * PAGE_SIZE;
         int end = Math.min(start + PAGE_SIZE, mailList.size());
         for (int i = start; i < end; i++) {
-            String rawMessage = "";
-            try {
-                rawMessage = new String(Files.readAllBytes(Paths.get(mailList.get(i))));
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-            MessageParser parser = new MessageParser();
+            String rawMessage = readRawMessage(mailList.get(i));
+            EmailParser parser = new EmailParser();
             parser.quickParse(rawMessage);
             String msgID = mailList.get(i).split("/")[2];
             msgID = msgID.substring(0, msgID.length() - 4);
@@ -125,7 +120,7 @@ public class ListEmails extends MainMenu {
             String subject = parser.getSubject();
             subject = formatString(subject, 30);
             String date = parser.getDate(SHORT_DATE_DISPLAY_FORMAT);
-            String attachment = (parser.hasAttachment() ? formatString("*", 10) : formatString("", 10));
+            String attachment = formatString(parser.hasAttachment() ? "*" : " ", 10);
 
             System.out.printf("[%d] %s| %s | %s | %s | %s |\n", i % 10, readStatus, sender, subject, date,
                     attachment);
@@ -145,6 +140,15 @@ public class ListEmails extends MainMenu {
         showOptions(options);
     }
 
+    private String readRawMessage(String filePath) {
+        try {
+            return new String(Files.readAllBytes(Paths.get(filePath)));
+        } catch (IOException e) {
+            e.printStackTrace();
+            return "";
+        }
+    }
+
     private void loadEmails() {
         Path path = mailbox.getMailboxDirectory();
         try {
@@ -152,7 +156,7 @@ public class ListEmails extends MainMenu {
                     .map(Path::toString)
                     .collect(Collectors.toList());
         } catch (IOException e) {
-            System.out.println(ANSI_TEXT_RED + "[ERROR] Error in listing emails." + ANSI_RESET);
+            displayErrorMessage("Error in listing emails.");
             e.printStackTrace();
         }
 
@@ -173,14 +177,18 @@ public class ListEmails extends MainMenu {
                 nextPage();
                 return true;
             case "Q":
+            case "q":
                 return false;
             case "D":
-                deleteEmailHandler(mailList, currentPage);
+            case "d":
+                deleteEmail(mailList, currentPage);
                 return true;
             case "M":
+            case "m":
                 moveEmailHandler(mailList, currentPage);
                 return true;
             case "S":
+            case "s":
                 EmailSearcher emailSearcher = new EmailSearcher(mailList, inputHandler);
                 emailSearcher.search();
                 return true;
@@ -203,12 +211,10 @@ public class ListEmails extends MainMenu {
         if (isInvalidOptions(userInput))
             return;
 
-        int emailIndex = currentPage * PAGE_SIZE + Integer.parseInt(userInput);
-        if (emailIndex >= mailList.size() || Integer.parseInt(userInput) > 9 || Integer.parseInt(userInput) < 0) {
-            System.out.printf("%s[ERROR] Invalid email number!%s\n", ANSI_TEXT_RED, ANSI_RESET);
-            sleep(TIME_2_SECONDS);
+        if (!validateEmailIndex(userInput))
             return;
-        }
+
+        int emailIndex = currentPage * PAGE_SIZE + Integer.parseInt(userInput);
 
         String emailDirectory = mailList.get(emailIndex);
         ViewEmail emailViewer = new ViewEmail(emailDirectory, mailList, emailIndex, mailboxes, inputHandler);
@@ -225,6 +231,17 @@ public class ListEmails extends MainMenu {
         emailViewer.showEmail();
     }
 
+    private boolean validateEmailIndex(String userInput) {
+        if (isInvalidOptions(userInput))
+            return false;
+        int emailIndex = currentPage * PAGE_SIZE + Integer.parseInt(userInput);
+        if (emailIndex >= mailList.size() || Integer.parseInt(userInput) > 9 || Integer.parseInt(userInput) < 0) {
+            displayErrorMessage("Invalid email number!");
+            return false;
+        }
+        return true;
+    }
+
     private String formatString(String original, int length) {
         if (original == null)
             return formatString("null", length);
@@ -234,18 +251,24 @@ public class ListEmails extends MainMenu {
         return original.format("%-" + length + "." + length + "s", original);
     }
 
-    private void deleteEmailHandler(List<String> emailList, int currentPage) {
+    private void deleteEmail(List<String> emailList, int currentPage) {
         String userInput = inputHandler.dialog("Mail to delete: ");
+        if (!validateEmailIndex(userInput))
+            return;
+
         int emailIndex = currentPage * PAGE_SIZE + Integer.parseInt(userInput);
+
+        displayStatusMessage("Deleting email...");
+
         Path emailPath = Paths.get(emailList.get(emailIndex));
         try {
             Files.delete(emailPath);
         } catch (IOException e) {
-            System.out.println(ANSI_TEXT_RED + "[ERROR] Error in deleting email." + ANSI_RESET);
+            displayErrorMessage(userInput + " is not a valid email number.");
             e.printStackTrace();
         }
         emailList.remove(emailIndex);
-        System.out.printf("%s%s%s\n", ANSI_TEXT_GREEN, "Email removed.", ANSI_RESET);
+        displaySuccessMessage("Email deleted.");
         sleep(TIME_2_SECONDS);
     }
 
@@ -254,8 +277,9 @@ public class ListEmails extends MainMenu {
         int emailIndex = currentPage * PAGE_SIZE + Integer.parseInt(userInput);
         String emailDirectory = mailList.get(emailIndex);
         String destination = inputHandler.dialog("Destination mailbox: ");
+        displayStatusMessage("Moving email...");
         Mailbox.moveMailToFolder(emailDirectory, destination);
-        System.out.printf("%s%s%s\n", ANSI_TEXT_GREEN, "Email moved to " + destination + ".", ANSI_RESET);
+        displaySuccessMessage("Email moved to " + destination + ".");
         sleep(TIME_2_SECONDS);
         mailList.remove(emailIndex);
     }
@@ -271,8 +295,8 @@ public class ListEmails extends MainMenu {
                 Files.delete(attachmentCachePath);
             }
         } catch (IOException e) {
-            System.out.println(ANSI_TEXT_RED + "[ERROR] Error in deleting attachments cache." + ANSI_RESET);
             e.printStackTrace();
+            displayErrorMessage("Error in deleting attachments cache.");
         }
     }
 }
